Response latencies are measured in microseconds
NoSQL databases are capable on sub-millisecond latencies and YCSB
was only measuring times with millisecond granularity. This caused
transactions that took place in less than a second to be recorded
as taking 0 time to complete and as a result had the potential to
skew test results. This fix changes the timing granularity to
microseconds.