[jdbc] made scan operation unimplemented
It was determined that since SQL does not have relevant scan
functionality, that it should be made unimplemented in the jdbc-binding.
It is possible to simulate scan results in SQL but the performance
metrics would not be comparable to a database system with true scan
functionality. The results also could not gaurantee consistency across
certain YCSB property configurations.