Extend the dynamoDB client to support the HASH_AND_RANGE primary key type. The default primary key type remains exactly the same as before: HASH type. And the default behavior and processing logic also remains exactly the same as before.
What's added is the support to benchmark a primary key type called HASH_AND_RANGE which is one of two supported primary key types of DynamoDB. Knowing the performance characteristics of the HASH_AND_RANGE key is important for some users because that's the only practical way for them to get strongly consistent query across multiple items.

See documentation in conf/dynamodb.properties for more details.