[elasticsearch] require path.home in embedded mode
Today when running the Elasticsearch binding in embedded mode, a
temporary directory is used. This can be confusing for running usual
load-then-run style workflows because the temporary directory will not
be the same between runs. Instead, we should just require a directory
when running in embedded mode and this commit makes this the case.