Update README.md
https://github.com/brianfrankcooper/YCSB/issues/1212

The YCSB readme is as the following, which is a little misleading.Because we can't use '-p durability=false' to disable durability. Instead it worked when I used '-p durability=SKIP_WAL'. The readme should be written more clearly.

durability: Whether or not writes should be appended to the WAL. Bypassing the WAL can improve throughput but data cannot be recovered in the event of a crash. The default is true.