Fixed Postgresnosql read operation missing tuple.
This patch fixes two bugs in the Postgresnosql benchmark read operation.
First, the offset accessed from the ResultSet of a read operation were incorrect, the possitions 2 and 3 were accessed when it should be 1 and 2.
Secondly, after a read operation, the ResultSet.next() was moving the cursor without adding the result to the YCSB map.