Adding a binding for YSQL with 2 keys to accomodate the scan scenario {workloade}.
Summary:
Scan performance of YCSB is really bad with hashing partitioning.
Workloade tries to simulate a workload where we are querying all the
posts for a thread and they do that with choosing a random key in the
usertable and then reading a bunch of keys after that. This obviously is
bad for us.

Instead of simulating the workload in such a sub-optimal way, we have
used a composite key to indicate the {thread,post}. This organization
works well with us and we get good performance.

Reviewers: Karthik, Mihnea, Neha