[hbase] HBase server side value filtering for long SCAN operations
The Apache HBase users usually implement primary use cases that fit well with HBase's intended short-scan and random get/put focus which already can be tested with YCSB.

However we can see in the industry that more and more users are also having secondary use cases that are better suited for traditional data warehouses,
like longer scan operations with filtering and also aggregation. We wish to contribute to YCSB to enable the performance measurement of these secondary workloads.

First we would like to submit a patch about enabling optional server side value filtering for HBase. In the Core Workload, we have values randomly generated as 100 long byte
arrays. We would introduce some HBase2 driver specific parameters that would setup value filtering on the scan operation, comparing a randomly generated 100 long byte array
with the values in the scan results. The operator used for comparison should be configurable as well (like "lessOrEqual", "greater", etc).

We are not very sure if this should be a generic feature of the Core Workload or not. We would implement this first as an HBase2 specific option (as the operators supported by
HBase might be limited / too HBase specific), but we are open for suggestions.