[cloudspanner] set max sessions
Previously, using the Cloud Spanner client with `threadcount` above 400
would error out with:

```
java.lang.IllegalArgumentException: Min sessions(1024) must be <= max sessions(400)
```

This is because the default max session limit is set to 400. This patch
sets the max session limit to the number of threads, to allow using more
than 400 threads.