feat: make JDBC connection URL delimiter configurable
Make the JDBC connection URL delimiter configurable. The current
delimiter is semicolon ';', which breaks JDBC drivers that use
semicolons in the actual connection URL. The default delimiter remains
semicolon for all databases, except a list of databases that are known
to use semicolons in the connection URL. Users can always specifiy a
custom delimiter to work around any issues with the default delimiter.

Fixes #1458