fix python ycsb script classpath building
The current python ycsb script drops `runtime` dependencies from the classpath. The script uses mvn dependency:build-classpath with includeScope=compile. The dependency plugin defines compile as compile + provided deps, which means that if a driver has a transitive runtime dep, it will be ignored.

This PR changes the classpath building to use dependency:list goal to dump all of the dependencies and excludes any test deps and then builds the classpath from that